flask_server/requirements.txt
torch
torchvision
flask
waitress # for production
Pillow # for client test
requests # for client test
argparse # for testing flask server
